+++ gazebo/gui/model/ModelEditorPalette.hh
@@ -101,18 +101,6 @@ namespace gazebo
       /// \brief Qt callback when model editing is complete.
       private slots: void OnDone();
 
-      /// \brief Widget that display model properties.
-      private: QTreeWidget *modelTreeWidget;
-
-      /// \brief Model settings item in the tree widget.
-      private: QTreeWidgetItem *modelSettingsItem;
-
-      /// \brief Model parts item in the tree widget.
-      private: QTreeWidgetItem *modelItem;
-
-      /// \brief Plugin item in the tree widget.
-      private: QTreeWidgetItem *pluginItem;
-
       /// \brief Parts button group.
       private: QButtonGroup *partButtonGroup;