+++ Changelog.md
@@ -8,13 +8,19 @@
 1. C++11 support (C++11 compatible compiler is now required)
     * [Pull request #1340](https://bitbucket.org/osrf/gazebo/pull-request/1340)
 
+1. Implemented private data pointer for the ODEPhysicsPrivate class.
+    * [Pull request #1383](https://bitbucket.org/osrf/gazebo/pull-request/1383)
+
 1. Implemented private data pointer for the World class.
     * [Pull request #1383](https://bitbucket.org/osrf/gazebo/pull-request/1383)
+
+1. Implemented private data pointer for the Scene class.
+    * [Pull request #1385](https://bitbucket.org/osrf/gazebo/pull-request/1385)
 
 1. Added a events::Event::resetWorld event that is triggered when World::Reset is called.
     * [Pull request #1332](https://bitbucket.org/osrf/gazebo/pull-request/1332)
     * [Issue #1375](https://bitbucket.org/osrf/gazebo/issue/1375)
- 
+
 1. Fixed `math::Box::GetCenter` functionality.
     * [Pull request #1278](https://bitbucket.org/osrf/gazebo/pull-request/1278)
     * [Issue #1327](https://bitbucket.org/osrf/gazebo/issue/1327)
@@ -181,6 +187,9 @@
     1. Pick textures from the palette and assign in 3D view.
         * [Pull request #1368](https://bitbucket.org/osrf/gazebo/pull-request/1368)
 
+    1. Pick custom colors from the palette and assign in 3D view.
+        * [Pull request #1382](https://bitbucket.org/osrf/gazebo/pull-request/1382)
+
 1. Model editor updates
     1. Fix adding/removing event filters .
         * [Pull request #1279](https://bitbucket.org/osrf/gazebo/pull-request/1279)
@@ -198,7 +207,7 @@
 
 1. GUI publishes model selection information on ~/selection topic.
     * [Pull request #1318](https://bitbucket.org/osrf/gazebo/pull-request/1318)
-            
+
 ## Gazebo 4.0
 
 ### Gazebo 4.x.x (yyyy-mm-dd)